+++ tibiawikisql/__main__.py
@@ -1,11 +1,24 @@
+import sqlite3
 import time
 
 import click
+from colorama import init
 
-from tibiawikisql.parsers import *
+from tibiawikisql import api
+from tibiawikisql import schema
+from tibiawikisql.models import SpellParser
+from tibiawikisql.models.item import ItemParser
+from tibiawikisql.parsers import spells
 
 __version__ = "2.0.0"
 DATABASE_FILE = "tibia_database.db"
+
+init()
+
+
+def progress_bar(iterable, label, length):
+    return click.progressbar(iterable=iterable, length=length, label=label,fill_char="█", empty_char="░",
+                             show_pos=True,  bar_template='%(label)s [\33[33m%(bar)s\33[0m] %(info)s')
 
 
 @click.group(context_settings={'help_option_names': ['-h', '--help']})
@@ -13,57 +26,72 @@ DATABASE_FILE = "tibia_database.db"
 def cli():
     pass
 
+categories = {
+    "spells": {"category": "Spells", "parser": SpellParser, "schema": schema.Spell},
+    "items": {"category": "Items", "parser": ItemParser, "schema": schema.Item},
+    # "creatures": {"category": "Creatures"},
+    # "keys": {"category": "Keys"},
+    # "npcs": {"category": "NPCs"},
+    # "achievements": {"category": "Achievements"},
+    # "imbuements": {"category": "Imbuements"},
+    # "quests": {"category": "Quest Overview Pages"},
+}
+
 
 @cli.command(name="generate")
 @click.option('-s', '--skip-images', help="Skip fetching and loading images to the database.", is_flag=True)
 @click.option('-db', '--db-name', help="Name for the database file.", default=DATABASE_FILE)
 def generate(skip_images, db_name):
     """Generates a database file."""
-    start_time = time.time()
-    print("Running...")
-    con = database.init_database(db_name)
+    print("Connecting to database...")
+    conn = sqlite3.connect(db_name)
+    print("Creating schema...")
+    schema.create_tables(conn)
+    conn.execute("PRAGMA synchronous = OFF")
+    data_store = {}
+    get_articles("Deprecated", data_store)
 
-    common.fetch_deprecated_list()
+    for key, value in categories.items():
+        get_articles(value["category"], data_store, key)
 
-    spells.fetch_spells_list()
-    spells.fetch_spells(con)
+    print("Parsing articles...")
+    for key, value in categories.items():
+        titles = [a.title for a in data_store[key]]
+        parser = value["parser"]
+        _schema = value["schema"]
+        unparsed = []
+        start = time.perf_counter()
+        generator = api.get_articles(titles)
+        extra_data = []
+        with conn:
+            with progress_bar(generator, f"Parsing {key}", len(titles)) as bar:
+                for i, article in enumerate(bar):
+                    row = parser.from_article(article, extra_data)
+                    if row is not None:
+                        _schema.insert(conn, **row)
+                    else:
+                        unparsed.append(titles[i])
+            if unparsed:
+                print(f"\33[31m\tCould not parse {len(unparsed):,} articles.\033[0m")
+                print("\t-> \33[31m%s\033[0m" % '\033[0m,\33[31m'.join(unparsed))
+            if extra_data:
+                results = parser.insert_extra_data(extra_data, conn)
+                for table, count in results.items():
+                    print(f"\tInserted {count:,} rows to \33[94m{table}\033[0m.")
+            dt = (time.perf_counter() - start)
+            print(f"\33[32m\tParsed articles in {dt:.2f} seconds.\033[0m")
 
-    items.fetch_items_list()
-    items.fetch_items(con)
-    items.fetch_keys_list()
-    items.fetch_keys(con)
-
-    imbuements.fetch_imbuements_list()
-    imbuements.fetch_imbuements(con)
-
-    npcs.fetch_npc_list()
-    npcs.fetch_npcs(con)
-
-    creatures.fetch_creature_list()
-    creatures.fetch_creature(con)
-    creatures.fetch_drop_statistics(con)
-
-    houses.fetch_house_list()
-    houses.fetch_houses(con)
-
-    achievements.fetch_achievement_list()
-    achievements.fetch_achievements(con)
-
-    quests.fetch_quest_list()
-    quests.fetch_quests(con)
-
-    npcs.save_rashid_locations(con)
-    if not skip_images:
-        creatures.fetch_creature_images(con)
-        items.fetch_item_images(con)
-        npcs.fetch_npc_images(con)
-        spells.fetch_spell_images(con)
-        imbuements.fetch_imbuements_images(con)
-        map.save_maps(con)
-
-    database.set_database_info(con, __version__)
-
-    print(f"Done in {time.time()-start_time:.3f} seconds.")
+def get_articles(category, data_store, key=None):
+    if key is None:
+        key = category.lower()
+    print(f"Fetching articles in \33[94mCategory:{category}\033[0m...")
+    data_store[key] = []
+    start = time.perf_counter()
+    for article in api.get_category_members(category):
+        if article not in data_store.get("deprecated", []):
+            data_store[key].append(article)
+    dt = (time.perf_counter() - start)
+    print(f"\33[32m\tFound {len(data_store[key]):,} articles in {dt:.2f} seconds.\033[0m")
 
 
 if __name__ == "__main__":
