+++ tibiawikisql/models/__init__.py
@@ -1,0 +1,2 @@
+from . import abc
+from .spell import SpellParser
